Int (%d) --> unsigned long integer (%lu)

+++ PASchedules/Classes/Controllers/Table View Controllers/Compare Tab/PAComparisonViewController.m
@@ -201,15 +201,15 @@ static NSString * kPACommitmentIdentifier = @"Commitment";
 
 - (NSString *)tableView:(UITableView *)tableView titleForHeaderInSection:(NSInteger)section {
     if (section == PAComparisonTableViewSectionCourses) {
-        return [NSString stringWithFormat:@"%@ (%d)", NSStringFromComparisonSections(section), self.sharedCourses.count];
+        return [NSString stringWithFormat:@"%@ (%lu)", NSStringFromComparisonSections(section), (unsigned long)self.sharedCourses.count];
     }
     
     if (section == PAComparisonTableViewSectionTeachers) {
-        return [NSString stringWithFormat:@"%@ (%d)", NSStringFromComparisonSections(section), self.sharedTeachers.count];
+        return [NSString stringWithFormat:@"%@ (%lu)", NSStringFromComparisonSections(section), (unsigned long)self.sharedTeachers.count];
     }
     
     if (section == PAComparisonTableViewSectionCommitments) {
-        return [NSString stringWithFormat:@"%@ (%d)", NSStringFromComparisonSections(section), self.sharedCommitments.count];
+        return [NSString stringWithFormat:@"%@ (%lu)", NSStringFromComparisonSections(section), (unsigned long)self.sharedCommitments.count];
     }
